update some changes user heretence

+++ src/main/java/com/example/FleetFlow/repositories/ChauffeurRepository.java
@@ -1,13 +1,12 @@
 package com.example.FleetFlow.repositories;
 
-import com.example.FleetFlow.DTO.ChauffeurDTO;
 import com.example.FleetFlow.models.Chauffeur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 
-public interface ChauffeurRepository extends JpaRepository<Chauffeur, Integer> {
+public interface ChauffeurRepository extends JpaRepository<Chauffeur, Long> {
     List<Chauffeur> findByIsDisponibleTrue();
     List<Chauffeur> findByPermisTypeAndIsDisponible(String permisType, Boolean isDisponible);
